Move Number of Schools table code to table folder
